+++ CHANGELOG.md
@@ -11,11 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - [Minor] `.after()` callback that can run after a specific field finished execution.
-- [Minor] New size rules (lessThan, greaterThan, lessThanOrEquals, greaterThanOrEquals).
-- [Minor] New size rules (longerThan, shorterThan, longerThanOrEquals, shorterThanOrEquals, lengthEquals).
+- [Minor] New size rules (`lessThan`, `greaterThan`, `lessThanOrEquals`, `greaterThanOrEquals`, `numberEquals`, `numberNotEquals`).
+- [Minor] New size rules (`longerThan`, `shorterThan`, `longerThanOrEquals`, `shorterThanOrEquals`, `lengthEquals`, `lengthNotEquals`).
 
 ### Removed
-- [Major] Removed compound rules (`anyOf()`, `allOf()`, `noneOf()`).
+- [Major] Removed compound rules (`anyOf`, `allOf`, `noneOf`).
+- [Major] Removed general size rules (`smallerThan`, `largerThan`, `smallerThanOrEquals`, `largerThanOrEquals`, `sizeEquals`, `sizeNotEquals`).
 
 ## [6.3.4] - 2019-05-01