+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [7.1.2] - 2019-06-27
+### Added
+- [Minor] Add isEven rule.
+- [Minor] Add isOdd rule.
+
+### Changed
+- [Patch] Make result callback functions non-enumerable
+
 
 ## [7.1.1] - 2019-06-24